% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_ORs.R
\name{brm_ors}
\alias{brm_ors}
\title{Generate Odds Ratios from logit link brm model}
\usage{
brm_ors(fit, digits = 2, predlabs, ...)
}
\arguments{
\item{fit}{model fit by brm function}

\item{digits}{number of digits to round}

\item{predlabs}{optional vetor of labels for each predictor}
}
\value{
Tibble with 3 columns: Name, Prob, 95% CI (as string)
}
\description{
Generate Odds Ratios from logit link brm model
}
\examples{
 require(brms)
x = runif(100,0,10)
y = sample(c(0,1), size = 100, replace = T)
z = runif(100,0,10)
df = tibble(x = x,y = y)
brmfit = brm(y ~ x + z, data = df, family = bernoulli(link = "logit"))
brm_pval(brmfit)
}
