% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_prob.R
\name{brm_probs}
\alias{brm_probs}
\title{Generate Probabilities from logit link model}
\usage{
brm_probs(fit, digits = 2, predlabs, ...)
}
\arguments{
\item{fit}{model fit by brm function}

\item{digits}{number of digits to round}

\item{predlabs}{optional vetor of labels for each predictor}
}
\value{
Tibble with 3 columns: Name, Prob, 95% CI (as string)
}
\description{
Generate Probabilities from logit link model
}
\examples{
x = runif(100,0,10)
y = sample(c(0,1), size = 100, replace = T)
z = runif(100,0,10)
g = rep(1:4,each = 25)
df = tibble(x = x,y = y, z = z, g = g)
brmfit = brm(y ~ x + z, data = df, family = bernoulli(link = "logit"))
brm_pval(brmfit)
}
