% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_pval.R
\name{brm_pval}
\alias{brm_pval}
\title{Generate empirical P-Value from MCMC generated by brm function}
\usage{
brm_pval(fit)
}
\arguments{
\item{brmfit}{Model fit by brm function}
}
\value{
tibble with 2 columns containing Fixed Effects names and pvalues
}
\description{
Generate empirical P-Value from MCMC generated by brm function
}
\examples{
require(brms)
x = 1:100
y = x^2 + rnorm(length(x))
g = rep(1:4,each = 25)
df = tibble(x = x,y = y,g = g)

brmfit = brm(y ~ x^2 + (1|g), data = df)
brm_pval(brmfit)
}
