% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_tab.R
\name{brm_tab}
\alias{brm_tab}
\title{Generate Coefficient Significance tables from brm model}
\usage{
brm_tab(fit, digits = 2, predlabs, ...)
}
\arguments{
\item{fit}{model fit by brm function}

\item{digits}{number of digits to round}

\item{predlabs}{optional vetor of labels for each predictor}
}
\value{
Tibble with 4 columns: Name, Prob, 95% CI (as string), pvalue
}
\description{
Generate Coefficient Significance tables from brm model
}
\examples{
require(brms)
x = 1:100
y = x^2 + rnorm(length(x))
z = runif(length(x),0,length(x))
g = rep(1:4,each = 25)
df = tibble(x = x,y = y,z = z, g = g)
brmfit = brm(y ~ x^2 + z + (1|g), data = df)
brm_tab(brmfit)
}
