% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_up.R
\name{count_up}
\alias{count_up}
\title{Count number of unique non-missing observations in a vector}
\usage{
count_up(x)
}
\arguments{
\item{x}{Vector of values to tally up}
}
\value{
integer representing number of non-missing values in a vector
}
\description{
This function tallies up the number of unique observations
within a vector. Input vector can be comprised of any data type
}
\examples{
x = 1:10
count_up(x)

x[c(1,5,7)] = NA
count_up(x)

x = sample(1:5, size = 100, replace = T)
count_up(x)

y = x[x != 5]
count_up(y)
}
