% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_count.R
\name{date_count}
\alias{date_count}
\title{Count up Frequency between two dates}
\usage{
date_count(x, lower = "0001-01-01", upper = "4000-01-01", rel = F, incl = F)
}
\arguments{
\item{x}{vector of Date objects}

\item{rel}{logical for output as relative frequency}

\item{incl}{logical for inclusion of endpoints}

\item{start}{Start of time interval in YYYY-MM-DD format}

\item{end}{End of time interval in YYYY-MM-DD format}
}
\value{
integer for total number of observations between time period or
double for relative frequency
}
\description{
Counts up number of elements in x between the start and end
point

lower and upper default to Jan 1AD and Jan 4000AD, so one sided intervals only need one

Counts up number of elements in x between the start and end
point
}
\examples{
soon.tm
}
