% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_Yr.R
\name{extract_Yr}
\alias{extract_Yr}
\title{Create middle IQR string from numeric vector}
\usage{
extract_Yr(x, format = "\%Y-\%m-\%d", numeric = F)
}
\arguments{
\item{x}{character string}

\item{format}{Date string format}

\item{numeric}{Return Year as numeric}
}
\value{
Year as a character string
}
\description{
Create middle IQR string from numeric vector
}
\examples{
extract_Yr(Sys.Date())
extract_Yr("1989-03-12")
extract_Yr("031289", format = "\%m\%d\%y")
extract_Yr("031289", format = "\%m\%d\%y", numeric = T)
extract_Yr("12_1989_03", format = "\%d_\%Y_\%m")
extract_Yr("12_1989_03", format = "\%d_\%Y_\%m", numeric = T)
}
