% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_w_date.R
\name{file_w_date}
\alias{file_w_date}
\title{Create File path with Today's Date appended}
\usage{
file_w_date(dir, file, ext, date = T)
}
\arguments{
\item{dir}{path to directory}

\item{file}{filename}

\item{ext}{extension type}

\item{date}{logical for Y/N appended date}
}
\value{
Character String with the path to filename
}
\description{
used to paste strings together for file location.
Primary usage is for long directory names that you can store in a string, as
well as for automatically appending the date to a file name for intermittenly
run reports/etc.
}
\examples{
file_w_date("C:/Documents", "Test", "csv")
data(mtcars)
DataEx = head(mtcars)
write.csv(DataEx, file_w_date("C:/Documents", "Test", "csv"))
write.csv(DataEx, file_w_date("C:/Documents", "Test", "csv", date = F))
}
