% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailp_uniHR.R
\name{frailp_uniHR}
\alias{frailp_uniHR}
\title{Format single predictor frailtypack `frailtyPenal` models into Hazard Ratio
table format}
\usage{
frailp_uniHR(mod, digits = 2)
}
\arguments{
\item{mod}{frailtyPenal object}

\item{digits}{number of digits top format to}
}
\value{
tibble with Hazard Ratio, 95-CI, and p value
}
\description{
Function accepts a frailtypack model and generates a one row table
for binding together with other models in univariate table
}
\examples{
TBD
}
