% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailp_uniHR_mimp.R
\name{frailp_uniHR_mimp}
\alias{frailp_uniHR_mimp}
\title{Format single predictor frailtypack `frailtyPenal` models into Hazard Ratio
table format from multiple imputed data sets}
\usage{
frailp_uniHR_mimp(modList, digits = 2)
}
\arguments{
\item{modList}{list of frailtypack objects}

\item{digits}{number of digits top format to}
}
\value{
tibble with Hazard Ratio, 95-CI, and p value
}
\description{
Function accepts a list of frailtypack models and generates a
one row table for binding together with other models in univariate table
}
\examples{
TBD
}
