% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_list.R
\name{freq_list}
\alias{freq_list}
\title{freq_list}
\usage{
freq_list(Tab)
}
\arguments{
\item{tab}{table to format}
}
\value{
tibble formulated with all combos and freq
}
\description{
Reformats table/3D Array into frequency list
Currently usage requires table generated with raw variable names after attach(<data>)
}
\examples{
attach(mtcars)
View(freq_list(table(cyl,vs,am,gear,carb)))
View(freq_list(table(cyl,vs,am)))
detach(mtcars)
}
