% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ORs.R
\name{glm_ORs}
\alias{glm_ORs}
\title{Generate Odds Ratios CI Pval table from GLM}
\usage{
glm_ORs(fit)
}
\arguments{
\item{fit}{model fit by brm function}
}
\value{
Tibble with 5 columns: Name (char), OR, 2.5%, 97.%, p-val (doubles)
}
\description{
Generate Odds Ratios CI Pval table from GLM
}
\examples{
x = runif(100,0,10)
y = sample(c(0,1), size = 100, replace = T)
z = runif(100,0,10)
df = tibble(
x = x,
y = y,
z = z
)
glmfit = glm(y ~ x + z, data = df, family = bernoulli(link = "logit"))
glm_ORs(glmfit)
}
