% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Better_and.R
\name{\%&\%}
\alias{\%&\%}
\title{T & NA -> F Operator}
\usage{
left \%&\% right
}
\arguments{
\item{left}{Left side of operator}

\item{right}{right side of operator}
}
\value{
vector of logicals
}
\description{
Replicates `&` behavior, but returns F instead of NA when making comparisons
of <exists> & NA. NA & NA still returns missing when comparing equal length
vectors.

Left and Right sides must be same length, or one of them length 1.

Equal length vectors will return vector of point-wise comparisons between the two

If left or right is length 1, it compares it to the entire other vector, returning
a vector of logicals comparing to each place.
}
\examples{
coming soon
}
