% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Better_equal.R
\name{\%==\%}
\alias{\%==\%}
\title{T == NA -> F Operator}
\usage{
left \%==\% right
}
\arguments{
\item{left}{Left side of operator}

\item{right}{right side of operator}
}
\value{
vector of logicals
}
\description{
Replicates `==` behavior, but returns F instead of NA when making comparisons
of <exists> == NA. NA == NA still returns missing when comparing equal length
vectors.

Left and Right sides must be same length, or one of them length 1.

Equal length vectors will return vector of point-wise comparisons between the two

If left or right is length 1, it compares it to the entire other vector, returning
a vector of logicals comparing to each place.
}
\examples{
colors = c("red","blue")
Col_Samp1 = sample(c(colors,NA), size = 10, replace = T, prob = c(.5,.2,.3))
Col_Samp2 = sample(c(colors,NA), size = 10, replace = T, prob = c(.5,.2,.3))

View(t(rbind(Col_Samp1,Col_Samp2,
       Col_Samp1  ==  Col_Samp2,
       Col_Samp1 \%==\% Col_Samp2)))

View(t(rbind(Col_Samp1,Col_Samp2,
       'red'  ==  Col_Samp1,
       'red' \%==\% Col_Samp1)))

T  ==  NA
T \%==\% NA

NA  ==  NA
NA \%==\% NA
}
