% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_csv.R
\name{make_csv}
\alias{make_csv}
\title{write.csv with more useful Defaults}
\usage{
make_csv(data, file)
}
\arguments{
\item{data}{data frame to output to csv}

\item{file}{name of the file to write to}
}
\value{
file
}
\description{
Replicates functionality of write.csv but with default options
na = ", row.names = F, and quote = F

Currently the function just calls write.csv with desired default options set
}
\examples{
coming soon
}
