% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimp_pool_mv.R
\name{mimp_pool_mv}
\alias{mimp_pool_mv}
\title{Obtain pooled estimate for multiple imputation analyses with multiple parameters}
\usage{
mimp_pool_mv(mods, mod0)
}
\arguments{
\item{mods}{list of models corresponding to MI datasets}

\item{mod0}{single model for same data without imputed predictor}
}
\value{
Pvalue for f test for inclusion of MI predictor
}
\description{
Serious WIP missing full functionality. Will be replaced
}
\examples{
TBD

}
