% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimp_pool_uv.R
\name{mimp_pool_uv}
\alias{mimp_pool_uv}
\title{Obtain pooled estimate for multiple imputation analyses}
\usage{
mimp_pool_uv(q, u)
}
\arguments{
\item{q}{Vector of estimates obtained from MI}

\item{u}{Vector of SEs corresponding to each estimate}
}
\value{
list with pooled estimate and se, along with t statistic and pvalue
}
\description{
Obtain pooled estimate for multiple imputation analyses
}
\examples{
TBD

}
