% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidate.R
\name{multidate}
\alias{multidate}
\title{Read Dates from varying forms}
\usage{
multidate(data, formats = c("\%m\%d\%Y", "\%m-\%d-\%y", "\%m/\%d/\%Y"))
}
\arguments{
\item{data}{Data Set (for piping)}

\item{formats}{Character vector containing the various date formats}
}
\value{
single Date object
}
\description{
Useful for combining with apply when columns
have various date formats
}
\examples{
multidate("03121989")
multidate("03-12-89")
multidate("03/12/1989")

##Not Run
library(tidyverse)
date_list = list(c("03121989","03-12-89","03/12/1989"))
dates = lapply(date_list, multidate)[[1]]

##Can also adpat to different formats for apply
date_list = list(c("128903","12198903","03/-12/-1989"))
my_multidate = function(x){
multidate(data = x, formats = c("\%d\%y\%m","\%d\%Y\%m","\%m/-\%d/-\%Y"))
}
dates = lapply(date_list, my_multidate)[[1]]
}
