% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Count Missing}
\usage{
na_count(x)
}
\arguments{
\item{x}{Vector}
}
\value{
integer for number of missing
}
\description{
Count Missing
}
\examples{
x = 1:100
ind = sample(1:100, 10, replace = F)
x[ind] = NA

x = 1:200
na_count(x)

x[ind] = NA
na_count(x)

}
