% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_squeeze.R
\name{num_squeeze}
\alias{num_squeeze}
\title{Squeezing numeric values within specified range}
\usage{
num_squeeze(x, low = -Inf, high = Inf)
}
\arguments{
\item{x}{vector of numeric values}

\item{low}{lower limit of numeric range}

\item{high}{upper limit of numeric range}
}
\value{
numeric vector
}
\description{
Coerces numbers outside the specified range to the min or max
value. low and high arguments default to `-Inf` and `Inf`
Function will return NULL if lower
}
\examples{
x = -100:100

num_squeeze(x)
num_squeeze(x, low = 0)
num_squeeze(x, low = -1e3, high = 1e3)
num_squeeze(x, low = -2, high = -5) #returns NULL with warning
num_squeeze(x, low = 0, high = 0)
}
