% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfm_HRs.R
\name{parfm_HRs}
\alias{parfm_HRs}
\title{Generate Hazard ratio, CI,  Pval table from parfm}
\usage{
parfm_HRs(fit, pred = NA)
}
\arguments{
\item{fit}{model fit by parfm function}

\item{pred}{Optional Argument for models with single predictor}
}
\value{
Tibble with 5 columns: Name (char), HR, 95 CI, p-val (doubles)
}
\description{
Generates table with Hazard rate, 95 CI and pval for parfm models fixed effects
WIP name extraction via other methods, until then names for models with single predictor
will have blank in Name column
}
\examples{
data(kidney)

fit = parfm(Surv(time,status) ~ age + sex + disease,
             cluster = "id", data = kidney,
             dist = "weibull", frailty = "gamma")

parfm_HRs(fit)
}
