% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dec.R
\name{print_dec}
\alias{print_dec}
\title{Function to force print to desired number of decimal points}
\usage{
print_dec(x, digits = 2)
}
\arguments{
\item{x}{Numeric Value}

\item{digits}{NUmber of digits to force printed}
}
\value{
character representation of numeric value with desired number of floating points
}
\description{
Function to force print to desired number of decimal points
}
\examples{
TBD
}
