% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_McNemar.R
\name{pwr_mcnemar}
\alias{pwr_mcnemar}
\title{Sample Size and Power for McNemar's Test}
\usage{
pwr_mcnemar(p10, p01, alpha = 0.05, n, power)
}
\arguments{
\item{p10}{Probability of pre-test success and post-test failure. See
details.}

\item{p01}{Probability of pre-test failure and post-test success. See
details.}

\item{alpha}{Specified significance level. Defaults to 0.05.}

\item{n}{Sample size. Cannot be left blank if \code{power} is missing.}

\item{power}{Statistical power. Cannot be left blank if \code{n} is missing.}
}
\description{
Computes sample size or power for McNemar's test for paired nominal data.
}
\details{
H0: Both groups have the same success probability.
H1: The success probability is not equal between the Groups.

                   Post
            Success    Failure
    Success   p11        p10
Pre
    Failure   p01        p00
}
\examples{
pwr.mcnemar(0.05, 0.45, power = 0.8)
pwr.mcnemar(0.05, 0.45, n = 22)
}
\references{
github/bgreenwell For Code

Connor R. J. 1987. Sample size for testing differences in proportions for the
paired-sample design. Biometrics 43(1):207-211. page 209.
}
