% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_comb_intv.R
\name{str_comb_Intv}
\alias{str_comb_Intv}
\title{Create center with spread data summary for table}
\usage{
str_comb_Intv(x, fun = median, digits = 2, delim = ",",
  limits = c(0.25, 0.75))
}
\arguments{
\item{x}{Numeric Vector}

\item{fun}{Function for Central Measure (Defaults to median)}

\item{digits}{number of digits to round}

\item{delim}{delimiter of choice}

\item{limits}{numeric vector of length 2 for lower and upper limit, defaults to IQR}
}
\value{
Character String of the form \code{fun(x) <delim> (Q1,Q3)}
}
\description{
Create center with spread data summary for table
}
\examples{
x = seq(1,100)
str_comb_Intv(x,digits = 3)
str_comb_Intv(x, lower = 0.1, upper = 0.85)
str_comb_Intv(x^2,fun = median, delim = " - ")
str_comb_Intv(x^2,fun = mean, delim = "_")
}
