% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_comb_NA.R
\name{str_comb_NA}
\alias{str_comb_NA}
\title{Count Missing and \% of Total}
\usage{
str_comb_NA(x, digits = 1, out = c("percentage", "percent"),
  zero2dash = T)
}
\arguments{
\item{x}{Vector}

\item{digits}{Number of decimal points to round to}

\item{out}{return percent or percentage}

\item{zero2dash}{automatically change '0 (0)' to '-'}
}
\value{
N (\%) of Number missing
}
\description{
Count Missing and \% of Total
}
\examples{
x = 1:100
ind = sample(1:100, 10, replace = F)
x[ind] = NA
str_comb_NA(x, out = "percent")
str_comb_NA(x, out = "percentage")

x = 1:200
str_comb_NA(x)

x[ind] = NA
str_comb_NA(x)
}
