% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_comb_Prop.r
\name{str_comb_prop}
\alias{str_comb_prop}
\title{Create count with proportion as string}
\usage{
str_comb_prop(x, category = NULL, na.rm = T, digits = 1,
  out = c("percentage", "percent"), perc.disp = F, escape = F,
  zero2dash = T)
}
\arguments{
\item{x}{vector}

\item{category}{Category to count x matches (not required if x logical)}

\item{digits}{number of digits to round}

\item{out}{single character representing output for proportion}

\item{perc.disp}{logical for inclusion of \% sign}

\item{escape}{logical for inclusion of escape character for LaTeX tables}

\item{zero2dash}{logical for returning "-" instead of "0 (0)" for tables}
}
\value{
Character String of the form "N ( \% )"
}
\description{
Create count with proportion as string
}
\examples{
x = sample(1:3, size = 20, replace = T)
str_comb_prop(x, category = 2,digits = 3)
str_comb_prop(x, category = 1, out = "percent")
str_comb_prop(x == 1)
str_comb_prop(x \%in\% 1:3)
}
