% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_excel2R.R
\name{str_excel2R}
\alias{str_excel2R}
\title{Replace PI excel column names with R friendly names}
\usage{
str_excel2R(colname)
}
\arguments{
\item{colname}{character vector for column names in dataset}
}
\value{
character vector with R friendly column names
}
\description{
Replace PI excel column names with R friendly names
}
\examples{
Coming soon
}
