% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sentence.R
\name{str_sentence}
\alias{str_sentence}
\title{Modify sentence to Capitalize only first word}
\usage{
str_sentence(string, noCaps = NULL)
}
\arguments{
\item{string}{string to put in sentence form}

\item{noCaps}{Additional Words to not capitalize}
}
\value{
Character String with proper capitalization
}
\description{
Coerces everything to lower case, then capitalizes the first
word and every word after that isn't 2-3 letter commonly lowercase word.

"in", "is","on", "at", "to", "from", "by","and", "but",
 "or", "for", "nor", "a", "an" all kept lowercase by default, additional
 words can be supplied with noCaps argument
}
\examples{
string = "HelLO, thIS IS A SEnTance oF SpONgEBoB mOcKInG yOu"
str_sentense(string)
}
