% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJFun.R
\name{sum.na}
\alias{sum.na}
\title{na.rm functions for funs}
\usage{
\method{sum}{na}(x)
}
\arguments{
\item{x}{numeric vector used in summarise}
}
\value{
single numeric value based on function name
}
\description{
na.rm functions for funs
}
\examples{
set.seed(1234)
Tib = tibble(x = 1:100,
             y = (1:100)^2)
NA.indx = sample(1:100, size = 10, replace = FALSE)
NA.indy = sample(1:100, size = 10, replace = FALSE)
Tib[NA.indx,1] = NA
Tib[NA.indy,2] = NA
Tib \%>\% summarise(Q1x = q1.na(X),
                  meany= mean.na(y))
}
