% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_full.R
\name{summ_full}
\alias{summ_full}
\title{5-Number summary as character string}
\usage{
summ_full(x, fun = median, digits = 2, delim = ",")
}
\arguments{
\item{x}{Numeric Vector}

\item{fun}{Function for Central Measure (Defaults to median)}

\item{digits}{number of digits to round}

\item{delim}{delimiter of choice}
}
\value{
Character String of the form \code{fun(x) <delim> (Q1,Q3) [min,max]}
}
\description{
5-Number summary as character string
}
\examples{
x = seq(1,100)
summ_full(x,digits = 3)
summ_full(x, lower = 0.1, upper = 0.85)
summ_full(x^2,fun = median, delim = " - ")
summ_full(x^2,fun = mean, delim = "_")
}
