% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_na.R
\name{summ_na}
\alias{summ_na}
\title{Count Missing and \% of Total}
\usage{
summ_na(x, digits = 1, out = c("percentage", "percent"),
  zero2dash = T)
}
\arguments{
\item{x}{Vector}

\item{digits}{Number of decimal points to round to}

\item{out}{return percent or percentage}

\item{zero2dash}{automatically change '0 (0)' to '-'}
}
\value{
N (\%) of Number missing
}
\description{
Count Missing and \% of Total
}
\examples{
x = 1:100
ind = sample(1:100, 10, replace = F)
x[ind] = NA
summ_na(x, out = "percent")
summ_na(x, out = "percentage")

x = 1:200
summ_na(x)

x[ind] = NA
summ_na(x)
}
