% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_swap.R
\name{temp_swap}
\alias{temp_swap}
\title{Temperature Transformation}
\usage{
temp_swap(faren = NA, cels = NA)
}
\arguments{
\item{faren}{Temperature in Farenheit}

\item{cels}{Temperature in Celsius}
}
\value{
vector of logicals
}
\description{
Accepts different combinations of temperatures in farenheit and Celsius
}
\examples{
coming soon

}
